pyproject.toml
[project]
name = "fortune-teller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.2",
    "fastmcp>=2.13.3",
    "py-clob-client>=0.31.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
